master2/METHODOLOGIE_TESTS/5_TD/methodologie-du-test/session_2/requirements.txt
pytest
pytest-cov
pytest-mock
python-dotenv
Flask
